+++ docs/src/app/layout.tsx
@@ -5,11 +5,10 @@ import { cn } from "@/lib/cn";
 import "@/styles/globals.css";
 import { Analytics } from "@vercel/analytics/react";
 import type { Metadata, Viewport } from "next";
-import { ThemeProvider } from "next-themes";
-import { ViewTransitions } from "next-view-transitions";
-
 import { Inter, JetBrains_Mono } from "next/font/google";
 import localFont from "next/font/local";
+import { ThemeProvider } from "next-themes";
+import { ViewTransitions } from "next-view-transitions";
 
 const fontSans = Inter({
   subsets: ["latin"],