+++ packages/core/src/index.ts
@@ -1,33 +1,90 @@
-import type { StandardSchemaDictionary, StandardSchemaV1 } from "./standard";
-import { parseWithDictionary } from "./standard";
+/**
+ * This is the core package of t3-env.
+ * It contains the `createEnv` function that you can use to create your schema.
+ * @module
+ */
+import type { StandardSchemaDictionary, StandardSchemaV1 } from "./standard.ts";
+import { ensureSynchronous, parseWithDictionary } from "./standard.ts";
 
-export type { StandardSchemaV1, StandardSchemaDictionary };
+export type {
+  /**
+   * The Standard Schema Interface
+   * @see https://github.com/standard-schema/standard-schema
+   * @internal
+   */
+  StandardSchemaV1,
+  /**
+   * A record with values being Standard Schema validators
+   * @see https://github.com/standard-schema/standard-schema
+   * @internal
+   */
+  StandardSchemaDictionary,
+};
 
-export type ErrorMessage<T extends string> = T;
-export type Simplify<T> = {
+/**
+ * Symbol for indicating type errors
+ * @internal
+ */
+type ErrorMessage<T extends string> = T;
+
+/**
+ * Simplify a type
+ * @internal
+ */
+type Simplify<T> = {
   [P in keyof T]: T[P];
 } & {};
 
+/**
+ * Get the keys of the possibly undefined values
+ * @internal
+ */
+type PossiblyUndefinedKeys<T> = {
+  [K in keyof T]: undefined extends T[K] ? K : never;
+}[keyof T];
+
+/**
+ * Make the keys of the type possibly undefined
+ * @internal
+ */
+type UndefinedOptional<T> = Partial<Pick<T, PossiblyUndefinedKeys<T>>> &
+  Omit<T, PossiblyUndefinedKeys<T>>;
+
+/**
+ * Make the keys of the type impossible
+ * @internal
+ */
 // biome-ignore lint/suspicious/noExplicitAny: <explanation>
 type Impossible<T extends Record<string, any>> = Partial<
   Record<keyof T, never>
 >;
 
-type UnReadonlyObject<T> = T extends Readonly<infer U> ? U : T;
+/**
+ * Reverse a Readonly object to be mutable
+ * @internal
+ */
+type Mutable<T> = T extends Readonly<infer U> ? U : T;
 
+/**
+ * Reduce an array of records to a single object where later keys override earlier ones
+ * @internal
+ */
 type Reduce<
-  TArr extends Array<Record<string, unknown>>,
-  TAcc = {},
+  TArr extends Record<string, unknown>[],
+  TAcc = object,
 > = TArr extends []
   ? TAcc
   : TArr extends [infer Head, ...infer Tail]
-    ? Tail extends Array<Record<string, unknown>>
-      ? Head & Reduce<Tail, TAcc>
+    ? Tail extends Record<string, unknown>[]
+      ? Mutable<Head> & Omit<Reduce<Tail, TAcc>, keyof Head>
       : never
     : never;
 
+/**
+ * The options that can be passed to the `createEnv` function.
+ */
 export interface BaseOptions<
-  TShared extends Record<string, StandardSchemaV1>,
+  TShared extends StandardSchemaDictionary,
   TExtends extends Array<Record<string, unknown>>,
 > {
   /**
@@ -81,8 +138,13 @@ export interface BaseOptions<
   emptyStringAsUndefined?: boolean;
 }
 
+/**
+ * Using this interface doesn't validate all environment variables are specified
+ * in the `runtimeEnv` object. You may want to use `StrictOptions` instead if
+ * your framework performs static analysis and tree-shakes unused variables.
+ */
 export interface LooseOptions<
-  TShared extends Record<string, StandardSchemaV1>,
+  TShared extends StandardSchemaDictionary,
   TExtends extends Array<Record<string, unknown>>,
 > extends BaseOptions<TShared, TExtends> {
   runtimeEnvStrict?: never;
@@ -95,11 +157,17 @@ export interface LooseOptions<
   runtimeEnv: Record<string, string | boolean | number | undefined>;
 }
 
+/**
+ * Using this interface validates all environment variables are specified
+ * in the `runtimeEnv` object. If you miss one, you'll get a type error. Useful
+ * if you want to make sure all environment variables are set for frameworks that
+ * perform static analysis and tree-shakes unused variables.
+ */
 export interface StrictOptions<
   TPrefix extends string | undefined,
-  TServer extends Record<string, StandardSchemaV1>,
-  TClient extends Record<string, StandardSchemaV1>,
-  TShared extends Record<string, StandardSchemaV1>,
+  TServer extends StandardSchemaDictionary,
+  TClient extends StandardSchemaDictionary,
+  TShared extends StandardSchemaDictionary,
   TExtends extends Array<Record<string, unknown>>,
 > extends BaseOptions<TShared, TExtends> {
   /**
@@ -129,9 +197,15 @@ export interface StrictOptions<
   runtimeEnv?: never;
 }
 
+/**
+ * This interface is used to define the client-side environment variables.
+ * It's used in conjunction with the `clientPrefix` option to ensure
+ * that all client-side variables are prefixed with the same string.
+ * Common examples of prefixes are `NEXT_PUBLIC_`, `NUXT_PUBLIC` or `PUBLIC_`.
+ */
 export interface ClientOptions<
   TPrefix extends string | undefined,
-  TClient extends Record<string, StandardSchemaV1>,
+  TClient extends StandardSchemaDictionary,
 > {
   /**
    * The prefix that client-side variables must have. This is enforced both at
@@ -152,9 +226,13 @@ export interface ClientOptions<
   }>;
 }
 
+/**
+ * This interface is used to define the schema for your
+ * server-side environment variables.
+ */
 export interface ServerOptions<
   TPrefix extends string | undefined,
-  TServer extends Record<string, StandardSchemaV1>,
+  TServer extends StandardSchemaDictionary,
 > {
   /**
    * Specify your server-side environment variables schema here. This way you can ensure the app isn't
@@ -173,10 +251,26 @@ export interface ServerOptions<
   }>;
 }
 
+export interface CreateSchemaOptions<
+  TServer extends StandardSchemaDictionary,
+  TClient extends StandardSchemaDictionary,
+  TShared extends StandardSchemaDictionary,
+  TFinalSchema extends StandardSchemaV1<{}, {}>,
+> {
+  /**
+   * A custom function to combine the schemas.
+   * Can be used to add further refinement or transformation.
+   */
+  createFinalSchema?: (
+    shape: TServer & TClient & TShared,
+    isServer: boolean,
+  ) => TFinalSchema;
+}
+
 export type ServerClientOptions<
   TPrefix extends string | undefined,
-  TServer extends Record<string, StandardSchemaV1>,
-  TClient extends Record<string, StandardSchemaV1>,
+  TServer extends StandardSchemaDictionary,
+  TClient extends StandardSchemaDictionary,
 > =
   | (ClientOptions<TPrefix, TClient> & ServerOptions<TPrefix, TServer>)
   | (ServerOptions<TPrefix, TServer> & Impossible<ClientOptions<never, never>>)
@@ -184,45 +278,60 @@ export type ServerClientOptions<
 
 export type EnvOptions<
   TPrefix extends string | undefined,
-  TServer extends Record<string, StandardSchemaV1>,
-  TClient extends Record<string, StandardSchemaV1>,
-  TShared extends Record<string, StandardSchemaV1>,
+  TServer extends StandardSchemaDictionary,
+  TClient extends StandardSchemaDictionary,
+  TShared extends StandardSchemaDictionary,
   TExtends extends Array<Record<string, unknown>>,
-> =
+  TFinalSchema extends StandardSchemaV1<{}, {}>,
+> = (
   | (LooseOptions<TShared, TExtends> &
       ServerClientOptions<TPrefix, TServer, TClient>)
   | (StrictOptions<TPrefix, TServer, TClient, TShared, TExtends> &
-      ServerClientOptions<TPrefix, TServer, TClient>);
+      ServerClientOptions<TPrefix, TServer, TClient>)
+) &
+  CreateSchemaOptions<TServer, TClient, TShared, TFinalSchema>;
 
 type TPrefixFormat = string | undefined;
-type TServerFormat = Record<string, StandardSchemaV1>;
-type TClientFormat = Record<string, StandardSchemaV1>;
-type TSharedFormat = Record<string, StandardSchemaV1>;
+type TServerFormat = StandardSchemaDictionary;
+type TClientFormat = StandardSchemaDictionary;
+type TSharedFormat = StandardSchemaDictionary;
 type TExtendsFormat = Array<Record<string, unknown>>;
 
-export type CreateEnv<
+export type DefaultCombinedSchema<
   TServer extends TServerFormat,
   TClient extends TClientFormat,
   TShared extends TSharedFormat,
-  TExtends extends TExtendsFormat,
-> = Readonly<
-  Simplify<
-    StandardSchemaDictionary.InferOutput<TServer> &
-      StandardSchemaDictionary.InferOutput<TClient> &
-      StandardSchemaDictionary.InferOutput<TShared> &
-      UnReadonlyObject<Reduce<TExtends>>
+> = StandardSchemaV1<
+  {},
+  UndefinedOptional<
+    StandardSchemaDictionary.InferOutput<TServer & TClient & TShared>
   >
 >;
 
+export type CreateEnv<
+  TFinalSchema extends StandardSchemaV1<{}, {}>,
+  TExtends extends TExtendsFormat,
+> = Readonly<
+  Simplify<Reduce<[StandardSchemaV1.InferOutput<TFinalSchema>, ...TExtends]>>
+>;
+
+/**
+ * Create a new environment variable schema.
+ */
 export function createEnv<
   TPrefix extends TPrefixFormat,
   TServer extends TServerFormat = NonNullable<unknown>,
   TClient extends TClientFormat = NonNullable<unknown>,
   TShared extends TSharedFormat = NonNullable<unknown>,
   const TExtends extends TExtendsFormat = [],
+  TFinalSchema extends StandardSchemaV1<{}, {}> = DefaultCombinedSchema<
+    TServer,
+    TClient,
+    TShared
+  >,
 >(
-  opts: EnvOptions<TPrefix, TServer, TClient, TShared, TExtends>,
-): CreateEnv<TServer, TClient, TShared, TExtends> {
+  opts: EnvOptions<TPrefix, TServer, TClient, TShared, TExtends, TFinalSchema>,
+): CreateEnv<TFinalSchema, TExtends> {
   const runtimeEnv = opts.runtimeEnvStrict ?? opts.runtimeEnv ?? process.env;
 
   const emptyStringAsUndefined = opts.emptyStringAsUndefined ?? false;
@@ -244,7 +353,7 @@ export function createEnv<
   const isServer =
     opts.isServer ?? (typeof window === "undefined" || "Deno" in window);
 
-  const finalSchema = isServer
+  const finalSchemaShape = isServer
     ? {
         ..._server,
         ..._shared,
@@ -255,7 +364,13 @@ export function createEnv<
         ..._shared,
       };
 
-  const parsed = parseWithDictionary(finalSchema, runtimeEnv);
+  const parsed =
+    opts
+      .createFinalSchema?.(finalSchemaShape as never, isServer)
+      ["~standard"].validate(runtimeEnv) ??
+    parseWithDictionary(finalSchemaShape, runtimeEnv);
+
+  ensureSynchronous(parsed, "Validation must be synchronous");
 
   const onValidationError =
     opts.onValidationError ??
@@ -290,7 +405,7 @@ export function createEnv<
   const extendedObj = (opts.extends ?? []).reduce((acc, curr) => {
     return Object.assign(acc, curr);
   }, {});
-  const fullObj = Object.assign(parsed.value, extendedObj);
+  const fullObj = Object.assign(extendedObj, parsed.value);
 
   const env = new Proxy(fullObj, {
     get(target, prop) {