+++ packages/core/src/presets-valibot.ts
@@ -1,7 +1,12 @@
-import { url, optional, picklist, pipe, string } from "valibot";
-import type { StandardSchemaDictionary } from ".";
-import { createEnv } from ".";
+/**
+ * Presets for Valibot
+ * @see https://env.t3.gg/docs/customization#extending-presets
+ * @module
+ */
+import { optional, picklist, pipe, string, url } from "valibot";
+import { createEnv } from "./index.ts";
 import type {
+  CoolifyEnv,
   FlyEnv,
   NeonVercelEnv,
   NetlifyEnv,
@@ -10,6 +15,7 @@ import type {
   SupabaseVercelEnv,
   UploadThingEnv,
   UploadThingV6Env,
+  UpstashRedisEnv,
   VercelEnv,
 } from "./presets";
 
@@ -17,7 +23,7 @@ import type {
  * Vercel System Environment Variables
  * @see https://vercel.com/docs/projects/environment-variables/system-environment-variables#system-environment-variables
  */
-export const vercel = () =>
+export const vercel = (): Readonly<VercelEnv> =>
   createEnv({
     server: {
       VERCEL: optional(string()),
@@ -41,7 +47,7 @@ export const vercel = () =>
       VERCEL_GIT_COMMIT_AUTHOR_NAME: optional(string()),
       VERCEL_GIT_PREVIOUS_SHA: optional(string()),
       VERCEL_GIT_PULL_REQUEST_ID: optional(string()),
-    } satisfies StandardSchemaDictionary.Matching<VercelEnv>,
+    },
     runtimeEnv: process.env,
   });
 
@@ -49,7 +55,7 @@ export const vercel = () =>
  * Neon for Vercel Environment Variables
  * @see https://neon.tech/docs/guides/vercel-native-integration#environment-variables-set-by-the-integration
  */
-export const neonVercel = () =>
+export const neonVercel = (): Readonly<NeonVercelEnv> =>
   createEnv({
     server: {
       DATABASE_URL: string(),
@@ -67,7 +73,7 @@ export const neonVercel = () =>
       POSTGRES_DATABASE: optional(string()),
       POSTGRES_URL_NO_SSL: optional(pipe(string(), url())),
       POSTGRES_PRISMA_URL: optional(pipe(string(), url())),
-    } satisfies StandardSchemaDictionary.Matching<NeonVercelEnv>,
+    },
     runtimeEnv: process.env,
   });
 
@@ -98,22 +104,22 @@ export const supabaseVercel = () =>
 /**
  * @see https://v6.docs.uploadthing.com/getting-started/nuxt#add-env-variables
  */
-export const uploadthingV6 = () =>
+export const uploadthingV6 = (): Readonly<UploadThingV6Env> =>
   createEnv({
     server: {
       UPLOADTHING_TOKEN: string(),
-    } satisfies StandardSchemaDictionary.Matching<UploadThingV6Env>,
+    },
     runtimeEnv: process.env,
   });
 
 /**
  * @see https://docs.uploadthing.com/getting-started/appdir#add-env-variables
  */
-export const uploadthing = () =>
+export const uploadthing = (): Readonly<UploadThingEnv> =>
   createEnv({
     server: {
       UPLOADTHING_TOKEN: string(),
-    } satisfies StandardSchemaDictionary.Matching<UploadThingEnv>,
+    },
     runtimeEnv: process.env,
   });
 
@@ -121,7 +127,7 @@ export const uploadthing = () =>
  * Render System Environment Variables
  * @see https://docs.render.com/environment-variables#all-runtimes
  */
-export const render = () =>
+export const render = (): Readonly<RenderEnv> =>
   createEnv({
     server: {
       IS_PULL_REQUEST: optional(string()),
@@ -138,7 +144,7 @@ export const render = () =>
         picklist(["web", "pserv", "cron", "worker", "static"]),
       ),
       RENDER: optional(string()),
-    } satisfies StandardSchemaDictionary.Matching<RenderEnv>,
+    },
     runtimeEnv: process.env,
   });
 
@@ -146,7 +152,7 @@ export const render = () =>
  * Railway Environment Variables
  * @see https://docs.railway.app/reference/variables#railway-provided-variables
  */
-export const railway = () =>
+export const railway = (): Readonly<RailwayEnv> =>
   createEnv({
     server: {
       RAILWAY_PUBLIC_DOMAIN: optional(string()),
@@ -172,7 +178,7 @@ export const railway = () =>
       RAILWAY_GIT_REPO_NAME: optional(string()),
       RAILWAY_GIT_REPO_OWNER: optional(string()),
       RAILWAY_GIT_COMMIT_MESSAGE: optional(string()),
-    } satisfies StandardSchemaDictionary.Matching<RailwayEnv>,
+    },
     runtimeEnv: process.env,
   });
 
@@ -180,7 +186,7 @@ export const railway = () =>
  * Fly.io Environment Variables
  * @see https://fly.io/docs/machines/runtime-environment/#environment-variables
  */
-export const fly = () =>
+export const fly = (): Readonly<FlyEnv> =>
   createEnv({
     server: {
       FLY_APP_NAME: optional(string()),
@@ -194,7 +200,7 @@ export const fly = () =>
       FLY_PROCESS_GROUP: optional(string()),
       FLY_VM_MEMORY_MB: optional(string()),
       PRIMARY_REGION: optional(string()),
-    } satisfies StandardSchemaDictionary.Matching<FlyEnv>,
+    },
     runtimeEnv: process.env,
   });
 
@@ -202,7 +208,7 @@ export const fly = () =>
  * Netlify Environment Variables
  * @see https://docs.netlify.com/configure-builds/environment-variables
  */
-export const netlify = () =>
+export const netlify = (): Readonly<NetlifyEnv> =>
   createEnv({
     server: {
       NETLIFY: optional(string()),
@@ -218,6 +224,38 @@ export const netlify = () =>
       DEPLOY_ID: optional(string()),
       SITE_NAME: optional(string()),
       SITE_ID: optional(string()),
-    } satisfies StandardSchemaDictionary.Matching<NetlifyEnv>,
+    },
+    runtimeEnv: process.env,
+  });
+
+/**
+ * Upstash redis Environment Variables
+ * @see https://upstash.com/docs/redis/howto/connectwithupstashredis
+ */
+export const upstashRedis = (): Readonly<UpstashRedisEnv> =>
+  createEnv({
+    server: {
+      UPSTASH_REDIS_REST_URL: pipe(string(), url()),
+      UPSTASH_REDIS_REST_TOKEN: string(),
+    },
+    runtimeEnv: process.env,
+  });
+
+/**
+ * Coolify Environment Variables
+ * @see https://coolify.io/docs/knowledge-base/environment-variables#predefined-variables
+ */
+export const coolify = (): Readonly<CoolifyEnv> =>
+  createEnv({
+    server: {
+      COOLIFY_FQDN: optional(string()),
+      COOLIFY_URL: optional(string()),
+      COOLIFY_BRANCH: optional(string()),
+      COOLIFY_RESOURCE_UUID: optional(string()),
+      COOLIFY_CONTAINER_NAME: optional(string()),
+      SOURCE_COMMIT: optional(string()),
+      PORT: optional(string()),
+      HOST: optional(string()),
+    },
     runtimeEnv: process.env,
   });