+++ packages/core/src/presets-zod.ts
@@ -1,7 +1,12 @@
+/**
+ * Presets for Zod
+ * @see https://env.t3.gg/docs/customization#extending-presets
+ * @module
+ */
 import { z } from "zod";
-import type { StandardSchemaDictionary } from ".";
-import { createEnv } from ".";
+import { createEnv } from "./index.ts";
 import type {
+  CoolifyEnv,
   FlyEnv,
   NeonVercelEnv,
   NetlifyEnv,
@@ -10,14 +15,15 @@ import type {
   SupabaseVercelEnv,
   UploadThingEnv,
   UploadThingV6Env,
+  UpstashRedisEnv,
   VercelEnv,
-} from "./presets";
+} from "./presets.ts";
 
 /**
  * Vercel System Environment Variables
  * @see https://vercel.com/docs/projects/environment-variables/system-environment-variables#system-environment-variables
  */
-export const vercel = () =>
+export const vercel = (): Readonly<VercelEnv> =>
   createEnv({
     server: {
       VERCEL: z.string().optional(),
@@ -41,7 +47,7 @@ export const vercel = () =>
       VERCEL_GIT_COMMIT_AUTHOR_NAME: z.string().optional(),
       VERCEL_GIT_PREVIOUS_SHA: z.string().optional(),
       VERCEL_GIT_PULL_REQUEST_ID: z.string().optional(),
-    } satisfies StandardSchemaDictionary.Matching<VercelEnv>,
+    },
     runtimeEnv: process.env,
   });
 
@@ -49,7 +55,7 @@ export const vercel = () =>
  * Neon for Vercel Environment Variables
  * @see https://neon.tech/docs/guides/vercel-native-integration#environment-variables-set-by-the-integration
  */
-export const neonVercel = () =>
+export const neonVercel = (): Readonly<NeonVercelEnv> =>
   createEnv({
     server: {
       DATABASE_URL: z.string(),
@@ -67,7 +73,7 @@ export const neonVercel = () =>
       POSTGRES_DATABASE: z.string().optional(),
       POSTGRES_URL_NO_SSL: z.string().url().optional(),
       POSTGRES_PRISMA_URL: z.string().url().optional(),
-    } satisfies StandardSchemaDictionary.Matching<NeonVercelEnv>,
+    },
     runtimeEnv: process.env,
   });
 
@@ -98,22 +104,22 @@ export const supabaseVercel = () =>
 /**
  * @see https://v6.docs.uploadthing.com/getting-started/nuxt#add-env-variables
  */
-export const uploadthingV6 = () =>
+export const uploadthingV6 = (): Readonly<UploadThingV6Env> =>
   createEnv({
     server: {
       UPLOADTHING_TOKEN: z.string(),
-    } satisfies StandardSchemaDictionary.Matching<UploadThingV6Env>,
+    },
     runtimeEnv: process.env,
   });
 
 /**
  * @see https://docs.uploadthing.com/getting-started/appdir#add-env-variables
  */
-export const uploadthing = () =>
+export const uploadthing = (): Readonly<UploadThingEnv> =>
   createEnv({
     server: {
       UPLOADTHING_TOKEN: z.string(),
-    } satisfies StandardSchemaDictionary.Matching<UploadThingEnv>,
+    },
     runtimeEnv: process.env,
   });
 
@@ -121,7 +127,7 @@ export const uploadthing = () =>
  * Render System Environment Variables
  * @see https://docs.render.com/environment-variables#all-runtimes
  */
-export const render = () =>
+export const render = (): Readonly<RenderEnv> =>
   createEnv({
     server: {
       IS_PULL_REQUEST: z.string().optional(),
@@ -138,7 +144,7 @@ export const render = () =>
         .enum(["web", "pserv", "cron", "worker", "static"])
         .optional(),
       RENDER: z.string().optional(),
-    } satisfies StandardSchemaDictionary.Matching<RenderEnv>,
+    },
     runtimeEnv: process.env,
   });
 
@@ -146,7 +152,7 @@ export const render = () =>
  * Railway Environment Variables
  * @see https://docs.railway.app/reference/variables#railway-provided-variables
  */
-export const railway = () =>
+export const railway = (): Readonly<RailwayEnv> =>
   createEnv({
     server: {
       RAILWAY_PUBLIC_DOMAIN: z.string().optional(),
@@ -172,7 +178,7 @@ export const railway = () =>
       RAILWAY_GIT_REPO_NAME: z.string().optional(),
       RAILWAY_GIT_REPO_OWNER: z.string().optional(),
       RAILWAY_GIT_COMMIT_MESSAGE: z.string().optional(),
-    } satisfies StandardSchemaDictionary.Matching<RailwayEnv>,
+    },
     runtimeEnv: process.env,
   });
 
@@ -180,7 +186,7 @@ export const railway = () =>
  * Fly.io Environment Variables
  * @see https://fly.io/docs/machines/runtime-environment/#environment-variables
  */
-export const fly = () =>
+export const fly = (): Readonly<FlyEnv> =>
   createEnv({
     server: {
       FLY_APP_NAME: z.string().optional(),
@@ -194,7 +200,7 @@ export const fly = () =>
       FLY_PROCESS_GROUP: z.string().optional(),
       FLY_VM_MEMORY_MB: z.string().optional(),
       PRIMARY_REGION: z.string().optional(),
-    } satisfies StandardSchemaDictionary.Matching<FlyEnv>,
+    },
     runtimeEnv: process.env,
   });
 
@@ -202,7 +208,7 @@ export const fly = () =>
  * Netlify Environment Variables
  * @see https://docs.netlify.com/configure-builds/environment-variables
  */
-export const netlify = () =>
+export const netlify = (): Readonly<NetlifyEnv> =>
   createEnv({
     server: {
       NETLIFY: z.string().optional(),
@@ -218,6 +224,38 @@ export const netlify = () =>
       DEPLOY_ID: z.string().optional(),
       SITE_NAME: z.string().optional(),
       SITE_ID: z.string().optional(),
-    } satisfies StandardSchemaDictionary.Matching<NetlifyEnv>,
+    },
+    runtimeEnv: process.env,
+  });
+
+/**
+ * Upstash redis Environment Variables
+ * @see https://upstash.com/docs/redis/howto/connectwithupstashredis
+ */
+export const upstashRedis = (): Readonly<UpstashRedisEnv> =>
+  createEnv({
+    server: {
+      UPSTASH_REDIS_REST_URL: z.string().url(),
+      UPSTASH_REDIS_REST_TOKEN: z.string(),
+    },
+    runtimeEnv: process.env,
+  });
+
+/**
+ * Coolify Environment Variables
+ * @see https://coolify.io/docs/knowledge-base/environment-variables#predefined-variables
+ */
+export const coolify = (): Readonly<CoolifyEnv> =>
+  createEnv({
+    server: {
+      COOLIFY_FQDN: z.string().optional(),
+      COOLIFY_URL: z.string().optional(),
+      COOLIFY_BRANCH: z.string().optional(),
+      COOLIFY_RESOURCE_UUID: z.string().optional(),
+      COOLIFY_CONTAINER_NAME: z.string().optional(),
+      SOURCE_COMMIT: z.string().optional(),
+      PORT: z.string().optional(),
+      HOST: z.string().optional(),
+    },
     runtimeEnv: process.env,
   });