+++ packages/core/test/smoke-valibot.test.ts
@@ -1,9 +1,7 @@
-/// <reference types="bun" />
-import { describe, expect, spyOn, test } from "bun:test";
-import { expectTypeOf } from "expect-type";
-
 import * as v from "valibot";
+import { describe, expect, expectTypeOf, test, vi } from "vitest";
 import { createEnv } from "../src";
+import { uploadthing } from "../src/presets-valibot";
 
 function ignoreErrors(cb: () => void) {
   try {
@@ -452,7 +450,7 @@ describe("extending presets", () => {
       }>
     >();
 
-    const consoleError = spyOn(console, "error");
+    const consoleError = vi.spyOn(console, "error");
     expect(() => lazyCreateEnv()).toThrow("Invalid environment variables");
     expect(consoleError.mock.calls[0]).toEqual([
       "❌ Invalid environment variables:",
@@ -628,4 +626,188 @@ describe("extending presets", () => {
       globalThis.window = window;
     });
   });
+});
+
+describe("createFinalSchema", () => {
+  test("custom schema combiner", () => {
+    let receivedIsServer = false;
+    const env = createEnv({
+      server: {
+        SERVER_ENV: v.string(),
+      },
+      shared: {
+        SHARED_ENV: v.string(),
+      },
+      clientPrefix: "CLIENT_",
+      client: {
+        CLIENT_ENV: v.string(),
+      },
+      runtimeEnv: {
+        SERVER_ENV: "server",
+        SHARED_ENV: "shared",
+        CLIENT_ENV: "client",
+      },
+      createFinalSchema: (shape, isServer) => {
+        expectTypeOf(isServer).toEqualTypeOf<boolean>();
+        if (typeof isServer === "boolean") receivedIsServer = true;
+        return v.object(shape);
+      },
+    });
+    expectTypeOf(env).toEqualTypeOf<
+      Readonly<{
+        SERVER_ENV: string;
+        SHARED_ENV: string;
+        CLIENT_ENV: string;
+      }>
+    >();
+    expect(env).toMatchObject({
+      SERVER_ENV: "server",
+      SHARED_ENV: "shared",
+      CLIENT_ENV: "client",
+    });
+    expect(receivedIsServer).toBe(true);
+  });
+  test("schema combiner with further refinement", () => {
+    const env = createEnv({
+      server: {
+        SKIP_AUTH: v.optional(v.boolean()),
+        EMAIL: v.optional(v.pipe(v.string(), v.email())),
+        PASSWORD: v.optional(v.pipe(v.string(), v.minLength(1))),
+      },
+      runtimeEnv: {
+        SKIP_AUTH: true,
+      },
+      createFinalSchema: (shape) =>
+        v.pipe(
+          v.object(shape),
+          v.check((env) => env.SKIP_AUTH || !!(env.EMAIL && env.PASSWORD)),
+        ),
+    });
+    expectTypeOf(env).toEqualTypeOf<
+      Readonly<{
+        SKIP_AUTH?: boolean;
+        EMAIL?: string;
+        PASSWORD?: string;
+      }>
+    >();
+    expect(env).toMatchObject({ SKIP_AUTH: true });
+  });
+  test("schema combiner that changes the type", () => {
+    const env = createEnv({
+      server: {
+        SKIP_AUTH: v.optional(v.boolean()),
+        EMAIL: v.optional(v.pipe(v.string(), v.email())),
+        PASSWORD: v.optional(v.pipe(v.string(), v.minLength(1))),
+      },
+      runtimeEnv: {
+        SKIP_AUTH: true,
+      },
+      createFinalSchema: (shape) =>
+        v.pipe(
+          v.object(shape),
+          v.rawTransform(({ addIssue, dataset, NEVER }) => {
+            const env = dataset.value;
+            if (env.SKIP_AUTH) return { SKIP_AUTH: true } as const;
+            if (!env.EMAIL || !env.PASSWORD) {
+              addIssue({
+                message:
+                  "EMAIL and PASSWORD are required if SKIP_AUTH is false",
+              });
+              return NEVER;
+            }
+            return {
+              EMAIL: env.EMAIL,
+              PASSWORD: env.PASSWORD,
+            };
+          }),
+        ),
+    });
+    expectTypeOf(env).toEqualTypeOf<
+      Readonly<
+        | {
+            readonly SKIP_AUTH: true;
+            EMAIL?: undefined;
+            PASSWORD?: undefined;
+          }
+        | {
+            readonly SKIP_AUTH?: undefined;
+            EMAIL: string;
+            PASSWORD: string;
+          }
+      >
+    >();
+    expect(env).toMatchObject({ SKIP_AUTH: true });
+  });
+});
+test("empty 'extends' array should not cause type errors", () => {
+  const env = createEnv({
+    clientPrefix: "FOO_",
+    server: { BAR: v.string() },
+    client: { FOO_BAR: v.string() },
+    runtimeEnvStrict: {
+      BAR: "bar",
+      FOO_BAR: "foo",
+    },
+    extends: [],
+  });
+
+  expectTypeOf(env).toEqualTypeOf<
+    Readonly<{
+      BAR: string;
+      FOO_BAR: string;
+    }>
+  >();
+
+  expect(env).toMatchObject({
+    BAR: "bar",
+    FOO_BAR: "foo",
+  });
+});
+
+test("overriding preset env var", () => {
+  const preset = createEnv({
+    server: {
+      PRESET_ENV: v.string(),
+    },
+    runtimeEnv: { PRESET_ENV: "preset" },
+  });
+
+  const env = createEnv({
+    server: {
+      PRESET_ENV: v.union([
+        v.pipe(v.string(), v.transform(Number)),
+        v.number(),
+      ]),
+    },
+    extends: [preset],
+    runtimeEnv: { PRESET_ENV: 123 },
+  });
+
+  expectTypeOf(env).toEqualTypeOf<
+    Readonly<{
+      PRESET_ENV: number;
+    }>
+  >();
+  expect(env.PRESET_ENV).toBe(123);
+});
+
+test("with built-in preset", () => {
+  process.env.UPLOADTHING_TOKEN = "token";
+  const env = createEnv({
+    server: {
+      FOO: v.string(),
+    },
+    extends: [uploadthing()],
+    runtimeEnv: { FOO: "bar" },
+  });
+
+  expectTypeOf(env).toEqualTypeOf<
+    Readonly<{
+      FOO: string;
+      UPLOADTHING_TOKEN: string;
+    }>
+  >();
+
+  expect(env.FOO).toBe("bar");
+  expect(env.UPLOADTHING_TOKEN).toBe("token");
 });