+++ packages/nextjs/src/index.ts
@@ -1,6 +1,14 @@
+/**
+ * This is the nextjs package of t3-env.
+ * It contains the `createEnv` function that you can use to create your schema.
+ * @module
+ */
 import type {
   CreateEnv,
+  CreateSchemaOptions,
+  DefaultCombinedSchema,
   ServerClientOptions,
+  StandardSchemaDictionary,
   StandardSchemaV1,
   StrictOptions,
 } from "@t3-oss/env-core";
@@ -10,13 +18,15 @@ const CLIENT_PREFIX = "NEXT_PUBLIC_" as const;
 type ClientPrefix = typeof CLIENT_PREFIX;
 
 type Options<
-  TServer extends Record<string, StandardSchemaV1>,
+  TServer extends StandardSchemaDictionary,
   TClient extends Record<`${ClientPrefix}${string}`, StandardSchemaV1>,
-  TShared extends Record<string, StandardSchemaV1>,
+  TShared extends StandardSchemaDictionary,
   TExtends extends Array<Record<string, unknown>>,
+  TFinalSchema extends StandardSchemaV1<{}, {}>,
 > = Omit<
   StrictOptions<ClientPrefix, TServer, TClient, TShared, TExtends> &
-    ServerClientOptions<ClientPrefix, TServer, TClient>,
+    ServerClientOptions<ClientPrefix, TServer, TClient> &
+    CreateSchemaOptions<TServer, TClient, TShared, TFinalSchema>,
   "runtimeEnvStrict" | "runtimeEnv" | "clientPrefix"
 > &
   (
@@ -53,17 +63,25 @@ type Options<
       }
   );
 
+/**
+ * Create a new environment variable schema.
+ */
 export function createEnv<
-  TServer extends Record<string, StandardSchemaV1> = NonNullable<unknown>,
+  TServer extends StandardSchemaDictionary = NonNullable<unknown>,
   TClient extends Record<
     `${ClientPrefix}${string}`,
     StandardSchemaV1
   > = NonNullable<unknown>,
-  TShared extends Record<string, StandardSchemaV1> = NonNullable<unknown>,
+  TShared extends StandardSchemaDictionary = NonNullable<unknown>,
   const TExtends extends Array<Record<string, unknown>> = [],
+  TFinalSchema extends StandardSchemaV1<{}, {}> = DefaultCombinedSchema<
+    TServer,
+    TClient,
+    TShared
+  >,
 >(
-  opts: Options<TServer, TClient, TShared, TExtends>,
-): CreateEnv<TServer, TClient, TShared, TExtends> {
+  opts: Options<TServer, TClient, TShared, TExtends, TFinalSchema>,
+): CreateEnv<TFinalSchema, TExtends> {
   const client = typeof opts.client === "object" ? opts.client : {};
   const server = typeof opts.server === "object" ? opts.server : {};
   const shared = opts.shared;
@@ -75,7 +93,14 @@ export function createEnv<
         ...opts.experimental__runtimeEnv,
       };
 
-  return createEnvCore<ClientPrefix, TServer, TClient, TShared, TExtends>({
+  return createEnvCore<
+    ClientPrefix,
+    TServer,
+    TClient,
+    TShared,
+    TExtends,
+    TFinalSchema
+  >({
     ...opts,
     shared,
     client,