+++ scripts/replace-workspace-protocol.ts
@@ -1,17 +1,9 @@
 /// <reference types="bun-types" />
 export const MODULE = true;
 
-declare module "bun" {
-  interface BunFile {
-    // biome-ignore lint/suspicious/noExplicitAny: <explanation>
-    json(): Promise<any>;
-  }
-}
-
 /**
  * Hack to replace the workspace protocol with the actual version
  */
-
 const corePkg = await Bun.file("../core/package.json").json();
 const version = corePkg.version;